+++ routes/api/auth.js
@@ -1,0 +1,8 @@
+const router = require('express').Router();
+const db = require('../../models');
+const authController = require('../../controllers/authController.js');
+const bcrypt = require('bcryptjs');
+
+router.post('/signin', authController.signin);
+
+module.exports = router;
