+++ routes/index.js
@@ -1,7 +1,11 @@
 const router = require('express').Router();
 const apiRouterUser = require('./api/users.js');
+const apiRouterAuth = require('./api/auth.js');
 
-router.use('/auth', apiRouterUser);
+
+//router.use('/users', apiRouterUsers);
+router.use('/auth', apiRouterAuth);
+
 
 //app.post('/api/auth', controller.signin)
